test: add more case
