+++ packages/core/test/fixtures/app-with-serverless-trigger/src/config/config.default.ts
