+++ packages/core/test/fixtures/app-with-serverless-trigger/src/config/config.prod.ts
