+++ packages/core/test/fixtures/app-with-serverless-trigger/src/function/hello_tencent.ts
